O365-Tools/O365Create/pyproject.toml
# Usage: 
#   Normal Install: pip install -e .
#   Build: python -m build
# be sure setuptools and build are in latest version
# pip install --upgrade setuptools build

[project]
name = "O365Create"
version = "1.0.0"
authors = [
    {name = "Mag. Stefan Hagmann"},
]
description = "Create O365 Accounts Helper"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.8"
dependencies = [
    'datetime',
    'fsspec',
    'rich',
    'pandas',
    'questionary',
    'universal-startfile',
    'pyyaml',
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
